pyproject.toml
[project]
name = "agent"
version = "0.0.1"
description = "Agent to parallel fetch all dashboard data."
authors = [
    { name = "Bhaskar Dutta", email = "bhaskar.dutta@c-4analytics.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.2.6",
    "python-dotenv>=1.0.1",
    "pandas>=2.0.0",
    "snowflake-connector-python==3.13.2",
    "snowflake.core==1.0.4",
    "transitions>=0.9.0",
    "requests>=2.31.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "typing-extensions>=4.5.0",
    "pydantic>=2.0.0",
    "dspy-ai==2.6.13"
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph.templates.agent", "agent"]
[tool.setuptools.package-dir]
"langgraph.templates.agent" = "agent"
"agent" = "agent"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"
